Integrate changes from e2x server

+++ nbgrader/exchange/collect.py
@@ -124,7 +124,8 @@ class ExchangeCollect(Exchange):
                     self.log.info("Updating submission: {} {}".format(student_id, self.coursedir.assignment_id))
                     shutil.rmtree(dest_path)
                     self.log.info("Updating hashed_submission directory")
-                    shutil.rmtree(hashed_dest_path)
+                    if os.path.isdir(hashed_dest_path):
+                        shutil.rmtree(hashed_dest_path)
                 else:
                     self.log.info("Collecting submission: {} {}".format(student_id, self.coursedir.assignment_id))
                 self.do_copy(src_path, dest_path)
@@ -141,33 +142,43 @@ class ExchangeCollect(Exchange):
                     ))
 
             user_info = self.extract_user_info(dest_path, student_id)
-            user_infos.append(user_info)
+            if user_info is not None:
+                user_infos.append(user_info)
 
         #Create hashcode list            
-        self.log.info("Creating hashcode list")
-        csv_filename = "hashcode_list.csv"
-        html_filename = "hashcode_list.html"
-        hashcode_list_path = os.path.join(self.coursedir.root, self.coursedir.submitted_directory)
-        csv_file_dest_path = os.path.join(hashcode_list_path, csv_filename)
-        html_file_dest_path = os.path.join(hashcode_list_path, html_filename)
-        self.create_hashcode_list(user_infos, hashcode_list_path, filename=csv_filename)
+        if len(user_infos) > 0: 
+            self.log.info("Creating hashcode list")
+            csv_filename = "{}_{}_hashcode_list.csv".format(self.coursedir.course_id, self.coursedir.assignment_id)
+            html_filename = "{}_{}_hashcode_list.html".format(self.coursedir.course_id, self.coursedir.assignment_id)
+            hashcode_list_path = os.path.join(self.coursedir.root, self.coursedir.submitted_directory)
+            csv_file_dest_path = os.path.join(hashcode_list_path, csv_filename)
+            html_file_dest_path = os.path.join(hashcode_list_path, html_filename)
+            self.create_hashcode_list(user_infos, hashcode_list_path, filename=csv_filename)
 
-        # Load csv and create html
-        self.log.info("Creating {}".format(html_file_dest_path))
-        data = pd.read_csv(csv_file_dest_path) 
-        data.to_html(html_file_dest_path, justify='center')
+            # Load csv and create html
+            self.log.info("Creating {}".format(html_file_dest_path))
+            data = pd.read_csv(csv_file_dest_path) 
+            data.to_html(html_file_dest_path, justify='center')
+
+            # Copy hashcode_list to /tmp for the other graders
+            #tmp_html_file_dest_path = os.path.join('/tmp', html_filename)
+            #data.to_html(tmp_html_file_dest_path, justify='center')
+        else:
+            self.log.info("Userinfo not found, hashcode list is not generated")
 
     def extract_user_info(self, info_path, username):
         file_info_path = os.path.join(info_path, username+"_info.txt")
-        try: 
-            with open(file_info_path) as f:
-                user_info = f.readlines()
-        finally:  
-            f.close()
+        if os.path.isfile(file_info_path):
+            try: 
+                with open(file_info_path) as f:
+                    user_info = f.readlines()
+            finally:  
+                f.close()
 
-        user_info = [x.strip().split(': ')[1] for x in user_info]
-
-        return user_info
+            user_info = [x.strip().split(': ')[1] for x in user_info]
+            return user_info
+        else:
+            return
 
     def create_hashcode_list(self, user_infos, list_dest_path, filename='hashcode_list.csv'):
         username_field = "Username"
